refactor: rename untils to utils
